// projects/hr-smart/src/app/webservice/generated/hr-smart-service/model/trainingEntry.ts
/**
 * HR-Smart Logging-Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DepartementEnum } from './departementEnum';
import { ParticipantEntry } from './participantEntry';
import { TrainingStatusEnum } from './trainingStatusEnum';
import { UserEntry } from './userEntry';


export interface TrainingEntry { 
    id?: string;
    subject?: string;
    topic?: string;
    departement?: DepartementEnum;
    date?: string;
    urgency?: string;
    hours?: number;
    description?: string;
    status?: TrainingStatusEnum;
    moderator?: UserEntry;
    allParticipant?: Array<ParticipantEntry>;
}