// projects/hr-smart/src/app/webservice/generated/hr-smart-service/model/userEntry.ts
/**
 * HR-Smart Logging-Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ParticipantEntry } from './participantEntry';


export interface UserEntry { 
    code?: string;
    roleID?: string;
    avatar?: string;
    trainings?: Array<ParticipantEntry>;
}